- Refactoring of the IFilesCommands and IAsyncFilesCommand related interfaces. - Cleanup of the AsyncFilesServerClient with better use of RavenJObject and RavenJArray deserialization extensions. - Added: Extensions to deal with plain object serialization (WriteWithObjectAsync).

+++ Raven.Database/Server/RavenFS/Controllers/RdcController.cs
@@ -47,8 +47,8 @@ namespace Raven.Database.Server.RavenFS.Controllers
 
                 var stats = new RdcStats
                 {
-                    CurrentVersion = rdcVersion.CurrentVersion,
-                    MinimumCompatibleAppVersion = rdcVersion.MinimumCompatibleAppVersion
+                    CurrentVersion = (int)rdcVersion.CurrentVersion,
+                    MinimumCompatibleAppVersion = (int)rdcVersion.MinimumCompatibleAppVersion
                 };
 
                 return this.GetMessageWithObject(stats, HttpStatusCode.OK)
